Group matches by doc title instead of memory location
Some unit tests for matches were failing in CI because the value
of the id() function used to group the match was not predictable.

Instead of using the memory location returned by id(), we make
the "title" attribute on docs required, so that we can use it
to more stably group matches when sorting.