Move is_norm_eq to method on Match
Calling the is_norm_eq function to check if a match is trivial
in the constructor of Matches would result in substantially more
calls, since the extension phase creates and discards a large
number of matches.

For performance, we move the call to a method that gets called
once per match during the filtering phase of the pipeline.